auth_service/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
kombu==5.3.7
mysqlclient==2.2.4
pika==1.3.2
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
responses==0.25.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
uwsgi